[refactor] Fix bugs and issues that SonarCloud reported, and fix JavaDoc errors

+++ exist-core/src/main/java/org/exist/xquery/functions/fn/FunStringToCodepoints.java
@@ -97,7 +97,7 @@ public class FunStringToCodepoints extends BasicFunction {
                 final int supp = XMLChar.supplemental(ch, s.charAt(++i));
                 next = new IntegerValue(expression, supp);
             } else {
-                next = new IntegerValue(expression, (int) ch);
+                next = new IntegerValue(expression, ch);
             }
             codepoints.add(next);
         }
